visual-service/requirements.txt
fastapi
uvicorn
torch
torchvision
pydantic
python-dotenv
playwright
Pillow
